Refactor playoff advancement code into own file
